Print Counter
Adds the function to store the number of prints in the EEPROM . Is based
in the code of @iashwinprasada with a few changes to make it work
independently if you have or not enable the EEPROM Settings.

Also i did a few changes in the way of implement this function with the
lcd, now is ready to add this feature in different languages currently
in (English and Spanish) need to translate the new menu
(MSG_PRINT_COUNTER, MSG_PRINTS).

The new Gcodes are

M504 - Retrieve the total number of prints
M505 - Increase the print counter by 1
M506 - Reset the print counter

And you can check the number of prins with the lcd in:

Menu\Control\Print Counter